Fix shard concatenation.
For batch learners, the shards of the dataset must be concatenated.
Previously, the shards were concatenated end to end. Since each shard
represents overlapping windows, this means the concatenated result is
out of order. Thus splitting the data between test and train was not
happening as expected. This lead to incorrect results which will be
updated in a future commit.